Renamed AutoLogging files to have AL prefix
Added the AL prefix to files to make it easier to mix with other
applications.
