+++ src/index.ts
@@ -32,5 +32,5 @@ export * as IReact from "@hyperion/hyperion-react/src/IReact";
 export * as IReactDOM from "@hyperion/hyperion-react/src/IReactDOM";
 
 // hyperionAutoLogging
-export * as Surface from "@hyperion/hyperion-autologging/src/Surface";
+export * as ALSurface from "@hyperion/hyperion-autologging/src/ALSurface";
 export { Channel } from "@hyperion/hook/src/Channel";
